services/data-producer/requirements.txt
# kafka-python ==2.0.2
# pandas ==1.5.3
# numpy ==1.24.2
# python-dotenv ==1.0.0

psycopg2-binary==2.9.7
numpy==1.25.2
pandas==1.4.2
SQLAlchemy==1.4.49
kafka-python==2.0.2